refactor(terminal): extract dev-permissions settings link, gate Mac hint at call site
- Move MacPermissionsHint out of terminal-pane/ since it renders from Terminal, not a pane
- Extract openDeveloperPermissionsSettings into a shared lib so the banner button and the
  pty-connection toast share one settings-routing entry point
- Hoist IS_MAC check to the Terminal call site so non-Mac platforms skip the subtree and
  its store subscriptions entirely
- Gate render on persistedUIReady to avoid a hint flash before dismissal hydrates
- Drop the explicit getDefaultUIState() default; the field is optional in PersistedUIState
  and the slice already coalesces with `?? false`, matching sibling dismissable flags
