+++ src/renderer/src/components/Terminal.tsx
@@ -22,7 +22,8 @@ import {
 import { Button } from '@/components/ui/button'
 import TabBar from './tab-bar/TabBar'
 import TerminalPane from './terminal-pane/TerminalPane'
-import { MacPermissionsHint } from './terminal-pane/MacPermissionsHint'
+import { MacPermissionsHint } from './MacPermissionsHint'
+import { IS_MAC } from '@/lib/new-workspace'
 import {
   ORCA_EDITOR_REQUEST_FILE_CLOSE_EVENT,
   ORCA_EDITOR_SAVE_AND_CLOSE_EVENT,
@@ -1231,11 +1232,17 @@ function Terminal(): React.JSX.Element | null {
           — tab groups + terminal extend to the top of the window instead.
           The old summary label (workspace / active surface) is removed. */}
 
-      <MacPermissionsHint
-        activeView={activeView}
-        activeTabType={activeTabType}
-        activeWorktreeId={activeWorktreeId}
-      />
+      {/* Why: sits above the layout-switching block so the hint is visible
+          regardless of which workspace render branch (split-group surface or
+          legacy single-pane) ends up active. Non-Mac platforms skip the
+          subtree entirely so its store subscriptions never wire up. */}
+      {IS_MAC && (
+        <MacPermissionsHint
+          activeView={activeView}
+          activeTabType={activeTabType}
+          activeWorktreeId={activeWorktreeId}
+        />
+      )}
 
       {anyMountedWorktreeHasLayout ? (
         <div
