+++ src/shared/constants.ts
@@ -263,7 +263,6 @@ export function getDefaultUIState(): PersistedUIState {
     statusBarVisible: true,
     dismissedUpdateVersion: null,
     lastUpdateCheckAt: null,
-    terminalMacPermissionsHintDismissed: false,
     trustedOrcaHooks: {}
   }
 }
